refactor: modularize features into separate files
- Split core_ext.rb and object_ext.rb into individual feature modules
- Move spec/feature/ to spec/features/ and add comprehensive tests
- Organize features: formattable, hash_mashable, implicit_calculable,
  onigregexp_ext, string_ext
- Update rubocop config and formatters for new structure
