+++ .rubocop.yml
@@ -72,5 +72,8 @@ RSpec/ContextWording:
 RSpec/DescribeClass:
   Enabled: false
 
+RSpec/NestedGroups:
+  Max: 4
+
 Style/PerlBackrefs:
   Enabled: false
