+++ mrblib/tempfile_ext.rb
@@ -1,0 +1,8 @@
+class Tempfile
+  def close(real: false)
+    super()
+    delete if real
+
+    nil
+  end
+end
