+++ spec/array_size_spec.rb
